gui/requirements.txt
streamlit>=1.34
pandas>=2.2
paho-mqtt>=1.6.1
